test: add perf test
